backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
matplotlib
opencv-python-headless
python-multipart
python-dotenv
